requirements.txt
pandas
google-generativeai
pytz